platform-devops/fastapi/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
httpx==0.27.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
pydantic==2.9.2
pydantic-settings==2.5.2
slowapi==0.1.9
redis==5.1.1
structlog==24.4.0
prometheus-fastapi-instrumentator==7.0.0
opentelemetry-sdk==1.27.0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-exporter-otlp==1.27.0
python-dotenv==1.0.1
